pyproject.toml
[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
[tool.pysen.lint.source]
  excludes = [".venv/",".git/", ".pytest_cache/", ".python-version/","data","tmp_data/"]
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]

[tool.poetry]
name = "typography-generation"
version = "0.1.0"
description = ""
authors = ["shimoda-uec <shimoda-k@mm.inf.uec.ac.jp>"]
readme = "README.md"
packages = [{include = "typography_generation", from = "src"}]


[tool.poetry.dependencies]
python = "^3.9"
skia-python = "^87.5"
einops = "^0.6.1"
hydra-core = "^1.3.2"
logzero = "^1.7.0"
datasets = "^2.12.0"
torch = "^1.13"
scikit-learn = "^1.0"
pytest = "^7.3.1"
pillow = "9.0.1"
matplotlib = "3.5"
transformers = "4.30.2"
openpyxl = "^3.1.2"
tensorboard = "^2.14.1"
gcsfs = "^2023.9.2"
seam-carving = "^1.1.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
notebook = "^6.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
